src/requirements.txt
asgiref==3.4.1
astroid==2.8.6
azure-core==1.20.1
azure-storage-blob==12.9.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.0
Django==3.2.9
django-dotenv==1.4.2
idna==3.3
isodate==0.6.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.1.1
packageurl-python==0.9.6
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pycparser==2.21
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.6
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
